More Animation Speed Options
Google's AOSP values are limiting.  This should allow more options for advanced users to play with until they find the exact animation speed they are looking for. At the very least, this commit allows for very fast .25 animation speeds for impatient users that just want the phone to run as fast as possible!